backend/api/requirements.txt
# FastAPI API Server Requirements

fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.30.0

# If you need authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6

# Production server
gunicorn>=21.2.0
